Gets past and future weather data
If asked for future weather data it gets 30 predictions of what the wheather will be. It then makes some stats like mean, std, etc so that we can get a single output of those 30 models. It does this for every hour in our time range, with every variable we request. Creates some graphs of the most relevant data. Note that the prediction changes daily as you approach the requested date.